requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.4
django-haystack==3.2.1
elasticsearch==7.17.3
pyelasticsearch==1.4.1
python-decouple==3.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
